requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.2.1
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
django-extensions==3.2.3
django-phonenumber-field==8.0.0
executing==2.1.0
ipython==8.28.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
phonenumbers==8.13.47
pillow==11.0.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
psycopg2-binary~=2.9.9
uwsgi
dateparser
celery[redis]==5.4.0
redis==5.2.1
flower
drf-yasg
python-decouple
